Use beveled shape when determining if a swipe should activate a key.
Instead of toggling whenever a key is activated, toggle the first activated key, and then set subsequent activated keys to the state of that first key. Whenever a swipe leaves the key area (as determined by the union of the bounding box of all keys), treat it as a new swipe that hasn't activated any keys yet.